share gateway type in iac
